feat: Create or add an exisitng organization using createOrganization() function.
Refactor createOrganization function and its test.

This commit gathers the changes made to the createOrganization() function to make it more efficient and readable.

The above changes include:
- Renaming `createOrg` function to `createOrganization`
- Using data validation to check if organization and user id exist or not
- If they do not exist, generating them using indexeddb
- Refactoring the code to make it more readable by separating the long if conditions
- Setting organization and user id to form Element
- Broadcasting the newly-created changed with socket.io to all connected clients

It also includes changes made to the test for `createOrganization` by renaming it from `createOrg` to `createOrganization` and changing the `endEvent` from `createdOrg` to `createdOrganization`.